ci: Add CI jobs which is package isolation test for detecting issues (#806)
* ﻿Enable testing in true isolation

We are experiencing a hang in our repo where we are building API Validation assemblies in which we take your latest promoted package `com.unity.renderstreaming@3.1.0-exp.4` (not happening with `3.1.0-exp.3`). The script imports the package with the Windows Unity Editor for the version you have specified in the `unity` field of your package (i.e. `2020.3`).

We think it's because the use of `SynchronizationContext` are not properly cleaned up after and Unity still waits for them. Unfortunately this doesn't reproduce when we run the same commands on the command line but this PR should, in theory, reproduce the same hang on your repo as well.

Enabling testing in true isolation is considered a good practice since this way you should be able to spot also missing dependencies since we are loading the package in a empty project and it should compile without errors.

We were also wondering the reason you added that huge --timeout arg to utr if it's related to this issue but we were unable to find any job that exceeded reasonable timings.

* ﻿add job for testing on Windows without GPU

* hopefully this will reproduce the issue

(cherry picked from commit 2583fbf5f484f734eb316c8758dbd38c96233ffd)

* add job

* fix

* ﻿add dependencies

* fix

* ﻿fix

* fix

* fix bug

* fix

* fix issues

* update dependency

Co-authored-by: Mihai Popescu <57262907+mihai-unity@users.noreply.github.com>